requirements.txt
# Core dependencies
langchain>=0.0.267
langgraph>=0.0.15
openai>=1.0.0
langchain_openai>=0.0.5

# Vector database
faiss-cpu>=1.7.4
langchain_community>=0.0.10

# Data validation
pydantic>=2.4.0

# API
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pytest>=7.4.0
python-dotenv>=1.0.0

# Visualization (optional)
matplotlib>=3.7.2
plotly>=5.16.0

# Type checking
mypy>=1.6.0
types-requests>=2.31.0.1

# Development tools
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
pymysql>=1.1.0
redis>=4.5.1

# Vector databases
pinecone-client>=2.2.1
pymilvus>=2.2.11
weaviate-client>=3.22.1

# Configuration management
python-dotenv>=1.0.0
pyyaml>=6.0.0